Added sessions

+++ server/controllers/userController.js
@@ -1,6 +1,8 @@
 const mongoose = require('mongoose')
 const User = mongoose.model('User')
 const bcrypt = require('bcrypt')
+const session = require('express-session')
+
 
 module.exports = {
     find: (req, res) => {
@@ -26,6 +28,8 @@ module.exports = {
                 hashedPassword: hash,
             })
             const newUser = await User.create(userToCreate);
+            req.session.isLoggedIn = true;
+            req.session.email = req.body.email;
             res.json(newUser)
         } catch (error) {
             res.json(error)
@@ -33,14 +37,34 @@ module.exports = {
         }
     },
     updateOne: (req, res) => {
+        if (!req.session.userId) return next(new Error('you need to log in first'))
         User.updateOne({_id: req.body._id},req.body,{new:true, runValidators:true})
             .then(data => res.json(data))
             .catch(error =>res.json(error))
     },
     deleteOne: (req, res) => {
+        if (!req.session.userId) return next(new Error('you need to log in first'))
         User.deleteOne({_id: req.params.id})
             .then(data => res.json(data))
             .catch(error =>res.json(error))
     },
+
+    login: async (req, res, next) => {
+        try {
+            let userInfo = await User.findOne({email: req.body.email})
+            if (!userInfo) return next(new Error('That email doesn\'t exist in the database'))
+            let match = await bcrypt.compare(req.body.password, userInfo.hashedPassword)
+            if (!match) return next(new Error('incorrect password'))
+            req.session.userId = userInfo._id
+            res.json(userInfo)
+        } catch (error) {
+            res.json(error)
+            next(error)
+        }
+    },
+    logout: async (req, res, next) => {
+        if (req.session.id) req.session.destroy();
+        res.json({message: 'success'})
+    }
 
 }
